+++ demo/src/main/java/com/project/demo/DesenhoAnimado.java
@@ -6,11 +6,15 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
 import lombok.EqualsAndHashCode;
 
+
+
 @Entity
 @EqualsAndHashCode(callSuper = true)
 @DiscriminatorValue("desenhoAnimado")
 public class DesenhoAnimado extends ItemCultural {
 
+    @Id @GeneratedValue
+    private long id;
 
     private String nome;
     private String criador;
@@ -21,6 +25,14 @@ public class DesenhoAnimado extends ItemCultural {
 
     public DesenhoAnimado() {
         super();
+    }
+
+    public long getId() {
+        return id;
+    }
+
+    public void setId(long id) {
+        this.id = id;
     }
 
     public String getNome() {
