--- demo/src/main/java/com/project/demo/security/controller/HelloWorldController.java
@@ -1,14 +1,0 @@
-package com.project.demo.security.controller;
-
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class HelloWorldController {
-
-	@RequestMapping({ "/hello" })
-	public String firstPage() {
-		return "hi";
-	}
-
-}
