--- demo/src/main/java/com/project/demo/security/model/User.java
@@ -1,24 +1,0 @@
-package com.project.demo.security.model;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-/**
- * The UserInfo class represents a user entity in the application.
- * It is used to store user-related data such as name, password etc.
- */
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Entity
-@Table(name = "users")
-public class User {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-    private String username;
-    private String password;
-
-}
