--- demo/src/main/java/com/project/demo/security/repo/UserRepo.java
@@ -1,18 +1,0 @@
-package com.project.demo.security.repo;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import com.project.demo.security.model.User;
-
-import java.util.Optional;
-
-
-@Repository
-public interface UserRepo extends JpaRepository<User, Integer> {
-    /**
-     * findByName method is used to retrieve a user by their username.
-     * It returns an Optional of UserInfo, which will be empty if no user is found.
-     */
-    Optional<User> findByUsername(String username);
-}
