--- demo/src/main/java/com/project/demo/security/service/JwtService.java
@@ -1,104 +1,0 @@
-package com.project.demo.security.service;
-
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.SignatureAlgorithm;
-import io.jsonwebtoken.io.Decoders;
-import io.jsonwebtoken.security.Keys;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.stereotype.Component;
-
-import java.security.Key;
-import java.util.Date;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.function.Function;
-
-
-//JwtService is responsible for handling JWT (JSON Web Token) operations
-// such as token generation, extraction of claims, and token validation.
-
-@Component
-public class JwtService {
-
-    // Secret Key for signing the JWT. It should be kept private.
-    private static final String SECRET = "TmV3U2VjcmV0S2V5Rm9ySldUU2lnbmluZ1B1cnBvc2VzMTIzNDU2Nzg=\r\n" + "";
-
-
-    // Generates a JWT token for the given username.
-    public String generateToken(String username) {
-        // Prepare claims for the token
-        Map<String, Object> claims = new HashMap<>();
-
-        // Build JWT token with claims, subject, issued time, expiration time, and signing algorithm
-          // Token valid for 180 minutes
-        return Jwts.builder()
-                .setClaims(claims)
-                .setSubject(username)
-                .setIssuedAt(new Date(System.currentTimeMillis()))
-                .setExpiration(new Date(System.currentTimeMillis() + 1000 * 60 * 60 * 5)) 
-                .signWith(getSignKey(), SignatureAlgorithm.HS256).compact();
-    }
-
-
-    // Creates a signing key from the base64 encoded secret.
-    //returns a Key object for signing the JWT.
-    private Key getSignKey() {
-        // Decode the base64 encoded secret key and return a Key object
-        byte[] keyBytes = Decoders.BASE64.decode(SECRET);
-        return Keys.hmacShaKeyFor(keyBytes);
-    }
-
-
-    // Extracts the username from the JWT token.
-    //return -> The username contained in the token.
-    public String extractUsername(String token) {
-        // Extract and return the subject claim from the token
-        return extractClaim(token, Claims::getSubject);
-    }
-
-
-    // Extracts the expiration date from the JWT token.
-    //@return The expiration date of the token.
-    public Date extractExpiration(String token) {
-        // Extract and return the expiration claim from the token
-        return extractClaim(token, Claims::getExpiration);
-    }
-
-
-    // Extracts a specific claim from the JWT token.
-    // claimResolver A function to extract the claim.
-    // return-> The value of the specified claim.
-    private <T> T extractClaim(String token, Function<Claims, T> claimResolver) {
-        // Extract the specified claim using the provided function
-        final Claims claims = extractAllClaims(token);
-        return claimResolver.apply(claims);
-    }
-
-    //Extracts all claims from the JWT token.
-    //return-> Claims object containing all claims.
-    private Claims extractAllClaims(String token) {
-        // Parse and return all claims from the token
-        return Jwts.parserBuilder()
-                .setSigningKey(getSignKey())
-                .build().parseClaimsJws(token).getBody();
-    }
-
-
-    //Checks if the JWT token is expired.
-    //return-> True if the token is expired, false otherwise.
-    public Boolean isTokenExpired(String token) {
-        // Check if the token's expiration time is before the current time
-        return extractExpiration(token).before(new Date());
-    }
-
-    //Validates the JWT token against the UserDetails.
-    //return-> True if the token is valid, false otherwise.
-
-    public Boolean validateToken(String token, UserDetails userDetails) {
-        // Extract username from token and check if it matches UserDetails' username
-        final String username = extractUsername(token);
-        // Also check if the token is expired
-        return (username.equals(userDetails.getUsername()) && !isTokenExpired(token));
-    }
-}
