+++ demo/src/main/resources/application.properties
@@ -2,6 +2,3 @@ spring.datasource.url=${SPRING_DATASOURCE_URL}
 spring.datasource.username=${SPRING_DATASOURCE_USERNAME}
 spring.datasource.password=${SPRING_DATASOURCE_PASSWORD}
 spring.jpa.hibernate.ddl-auto=update
-logging.level.org.springframework.web=DEBUG
-spring.resources.static-locations=classpath:/static/
-spring.web.resources.add-mappings=true
