+++ demo/src/main/resources/static/documentarios.html
@@ -48,7 +48,7 @@
         <tbody id="corpoTabelaDocumentarios"> </tbody>
     </table>
 
-    <script src="documentarios-rest.js" type="application/javascript"></script>
+    <script src="documentarios-rest.js"></script>
     <script>
         const tabelaDocumentarios = document.querySelector('#tabelaDocumentarios');
         const tabelaFormulario = document.querySelector('#tabelaFormulario');
@@ -224,7 +224,6 @@
 
         function alterarDocumentario() {
             const dadosDocumentario = {
-                'item_type': 'DOCUMENTARIO', // Adicionado o tipo
                 'id': txtId.value,
                 'nome': txtNome.value,
                 'diretor': txtDiretor.value,
