--- demo/src/main/resources/static/html/login.html
@@ -1,28 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <title>Login</title>
-        <meta charset="UTF-8">
-        <link rel="stylesheet" href="https://unpkg.com/sakura.css/css/sakura.css" type="text/css">
-    </head>
-<body>
-  <h1>Login</h1>
-  <div class="input">
-  	<table>
-    	<tr>
-    		<td><label for="username">user name:</label></td>
-    		<td><input id="username"></td>
-    	</tr>
-    	<tr>
-    		<td><label for="password">password:</label></td>
-    		<td><input type="password" id="password"></td>
-    	</tr>
-    	<tr><td></td><td><button onclick="getToken()">login</button></td></tr>
-    	<tr><td></td><td><a href="cadastro.html">Criar novo usuário</a></td></tr>
-    </table>
-  </div>
-  <div id="message"></div>
-
-  <script src="/js/jwt-vanilla.js"></script>
-</body>
-</html>
