+++ demo/src/main/resources/static/livros-rest.js
