+++ demo/src/main/resources/static/series-rest.js
@@ -1,19 +1,9 @@
 async function asyncCriarSerie(dadosSerie, proxsucesso, proxerro) {
-    const token = localStorage.getItem('token');
-    if (!token) {
-        alert('Token não encontrado. Por favor, faça login novamente.');
-        window.location.replace('/html/login.html'); // Redireciona para a página de login
-        return;
-    }
-
     const URL = `/api/series`;
     const postRequest = {
         method: 'POST',
         body: JSON.stringify(dadosSerie),
-        headers: { 
-            'Content-type': 'application/json',
-            'Authorization': `Bearer ${token}` // Incluindo o token no cabeçalho
-        }
+        headers: { 'Content-type': 'application/json' }
     };
     fetch(URL, postRequest)
         .then(resposta => { if (!resposta.ok) throw Error(resposta.status); return resposta; })
@@ -23,13 +13,6 @@ async function asyncCriarSerie(dadosSerie, proxsucesso, proxerro) {
 }
 
 async function carregarSeries() {
-    const token = localStorage.getItem('token');
-    if (!token) {
-        alert('Token não encontrado. Por favor, faça login novamente.');
-        window.location.replace('/html/login.html'); // Redireciona para a página de login
-        return;
-    }
-
     asyncLerSerieById(
         series => {
             console.log("Séries carregadas:", series);
@@ -43,20 +26,8 @@ async function carregarSeries() {
 }
 
 async function asyncLerSeriesById(proxsucesso, proxerro) {
-    const token = localStorage.getItem('token');
-    if (!token) {
-        alert('Token não encontrado. Por favor, faça login novamente.');
-        window.location.replace('/html/login.html'); // Redireciona para a página de login
-        return;
-    }
-
     const URL = `/api/series`; // Endpoint para obter todas as séries
-    fetch(URL, {
-        method: 'GET',
-        headers: {
-            'Authorization': `Bearer ${token}` // Incluindo o token no cabeçalho
-        }
-    })
+    fetch(URL)
         .then(resposta => {
             if (!resposta.ok) throw Error(resposta.status);
             return resposta.json();
@@ -64,23 +35,12 @@ async function asyncLerSeriesById(proxsucesso, proxerro) {
         .then(jsonResponse => proxsucesso(jsonResponse))
         .catch(proxerro);
 }
-
 async function asyncAlterarSerie(dadosSerie, proxsucesso, proxerro) {
-    const token = localStorage.getItem('token');
-    if (!token) {
-        alert('Token não encontrado. Por favor, faça login novamente.');
-        window.location.replace('/html/login.html'); // Redireciona para a página de login
-        return;
-    }
-
     const URL = `/api/series/${dadosSerie.id}`;
     const putRequest = {
         method: 'PUT',
         body: JSON.stringify(dadosSerie),
-        headers: { 
-            'Content-type': 'application/json',
-            'Authorization': `Bearer ${token}` // Incluindo o token no cabeçalho
-        }
+        headers: { 'Content-type': 'application/json' }
     };
     fetch(URL, putRequest)
         .then(resposta => { if (!resposta.ok) throw Error(resposta.status); return resposta; })
@@ -96,20 +56,8 @@ async function asyncApagarSerie(id, proxsucesso, proxerro) {
         return;
     }
 
-    const token = localStorage.getItem('token');
-    if (!token) {
-        alert('Token não encontrado. Por favor, faça login novamente.');
-        window.location.replace('/html/login.html'); // Redireciona para a página de login
-        return;
-    }
-
     const URL = `/api/series/${id}`;
-    const deleteRequest = {
-        method: 'DELETE',
-        headers: { 
-            'Authorization': `Bearer ${token}` // Incluindo o token no cabeçalho
-        }
-    };
+    const deleteRequest = { method: 'DELETE' };
 
     fetch(URL, deleteRequest)
         .then(resposta => {
