+++ demo/src/main/resources/static/series.html
@@ -65,7 +65,7 @@
         <tbody id="corpoTabelaSeries"> </tbody>
     </table>
 
-    <script src="series-rest.js" type="application/javascript"></script>
+    <script src="series-rest.js"></script>
     <script>
         const tabelaSeries = document.querySelector('#tabelaSeries');
         const tabelaFormulario = document.querySelector('#tabelaFormulario');
